pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "posthog-otel-exporter"
version = "0.1.0"
description = "Export OpenTelemetry-compatible events to PostHog."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opentelemetry-sdk>=1.38.0",
    "posthog>=7.0.1",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.4.0",
    "pytest>=9.0.1",
    "ruff>=0.14.5",
]

[tool.hatch.build.targets.wheel]
only-include = ["src/posthog_otel_exporter"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""